Alternative approach
This is a total rewrite of the library, starting with an observation
from Andres that we could try to take advantage of the fact that we can
use non-standard `Rep` instances for GHC generics, so that it looks as
if `large-records` does in fact support GHC generics. This doesn't mean
it's free: we still need to provide custom _instances_ for all the
generic classes that `beam` uses, and there are quite a few of those.
However, this provides a much smoother integration for the user of the
library; there barely are differences now between using LR and not using
LR, as far as beam integration is concerned. All that is required is

```haskell
import Data.Record.Beam ()
```

Indeed, apart from instances, there literally is nothing to import.

This also enables us to reuse much more of beam internally, rather than
having to duplicate a lot of its functionality.

This rewrite offers all the functionality of the original library (the
same test cases are still present), and significantly more: the three
`beam` tutorials now all work with `large-records` (added test cases for
these). Moreover, combinations of large-records and regular tables in
a single database should now also be possible.